requirements.txt
python-twitter==3.2
MySQL-python==1.2.5
